Move snapshots out of utils
They are not utilities, it's a primitive model.
